# Conflicts:
#	src/components/DivisionReport/DetailedCallOutReport.tsx

+++ src/pages/api/admin/callouts.ts
@@ -1,6 +1,6 @@
 // Next.js API route support: https://nextjs.org/docs/api-routes/introduction
 import {Request, Response} from 'express';
-import {enforceAdminOnly, JwtPayload} from '../../../auth';
+import {getJwtTokenForAPI, JwtPayload} from '../../../auth';
 import deleteEmployeeCallOutApiHandler from '../../../lib/apiController/callouts/DELETE';
 import {editEmployeeCallOutApiHandler, getCallOutsApiHandler} from '../../../lib/apiController';
 
@@ -10,15 +10,21 @@ export default async function handler(req: Request, res: Response) {
     return getCallOutsApiHandler(req, res);
   }
 
-  // If not a GET request, then the user must be an admin
-  const token: JwtPayload | undefined | Response<any, Record<string, any>> | void =
-    await enforceAdminOnly(req, res);
-
   if (req.method === 'PUT') {
+    const token = await getJwtTokenForAPI(req, res);
+    if (!token) {
+      return;
+    }
     return editEmployeeCallOutApiHandler(req, res, token as JwtPayload);
   }
 
   if (req.method === 'DELETE') {
+    // Deletion remains admin-only.
+    const token = await getJwtTokenForAPI(req, res);
+    if (!token?.isAdmin) {
+      res.status(401).json({error: 'Unauthorized request'});
+      return;
+    }
     return deleteEmployeeCallOutApiHandler(req, res);
   }
   return res.status(405).json({error: 'Method not allowed'});
